% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayerhanck.R
\name{bayerhanck}
\alias{bayerhanck}
\title{Joint Test-Statistic for the Null of Non-Cointegration}
\usage{
bayerhanck(formula, data, lags = 1, trend = "const", test = "all")
}
\arguments{
\item{formula}{An object of class \code{\link[stats]{formula}} to describe the model.}

\item{data}{An optional data frame containing the variables in the model.}

\item{lags}{Number of lags to be included.}

\item{trend}{Type of deterministic component to be included, "none" for no deterministic,
"const" for a constant and "trend" for a constant plus trend.}

\item{test}{Selection of tests to choose from. Choices are either "ej", for \code{\link{englegranger}}
and \code{\link{johansen}}, or "all", for \code{\link{englegranger}}, \code{\link{johansen}},
\code{\link{banerjee}} and \code{\link{boswijk}}.}
}
\value{
\code{bayerhanck} returns an object of classes \code{"bh.test"} and \code{"list"}.

The function \code{summary} is used to print a summary, whereas the cumulative distribution
under the null hypothesis can be plotted with \code{plot}.
}
\description{
Produces a joint test-statistic for the null of non-cointegration, aggregating
various cointegration tests.
}
\examples{
data("mts-examples", package="MTS")
bayerhanck(sp ~ ibm + ko, data = ibmspko)
}
\references{
Bayer, C. and Hanck, C. (2009), Combining Non-Cointegration tests, METEOR RM 09/012, University of Maastricht.
}
